add widgets for each resources

+++ database/migrations/2023_12_28_135021_create_purchasing_expenses_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('item_name');
             $table->integer('quantity')->nullable();
-            $table->integer('unit_price')->nullable();
-            $table->integer('total_cost');
+            $table->bigInteger('unit_price')->nullable();
+            $table->bigInteger('total_cost');
             $table->longText('description')->nullable();
             $table->timestamps();
         });
